# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/events/admin": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "관리자용 실시간 이벤트 스트림",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "F. Events (SSE)"
                ],
                "summary": "Admin SSE Stream",
                "responses": {
                    "200": {
                        "description": "SSE Stream"
                    }
                }
            }
        },
        "/api/v1/events/track/{trackId}": {
            "get": {
                "security": [
                    {
                        "TrackAuth": []
                    }
                ],
                "description": "트랙 진행자용 실시간 이벤트 스트림",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "F. Events (SSE)"
                ],
                "summary": "Track SSE Stream",
                "parameters": [
                    {
                        "type": "string",
                        "description": "트랙 ID",
                        "name": "trackId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE Stream"
                    }
                }
            }
        },
...
        "/device-registrations": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "관리자가 등록되었거나 대기 중인 기기 목록을 확인한다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "A. Auth \u0026 Device Trust"
                ],
                "summary": "기기 등록 목록 조회",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_infrastructure_web.DeviceRegistration"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "기기가 서버에 등록을 요청한다. 이후 관리자의 승인 대기.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "A. Auth \u0026 Device Trust"
                ],
                "summary": "기기 등록 요청 (최초 앱 실행 시)",
                "parameters": [
                    {
                        "description": "등록 정보",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.DeviceRegistrationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.DeviceRegistration"
                        }
                    }
                }
            }
        },
...
        "/device-registrations/{id}/approve": {
            "post": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "PENDING 상태인 기기를 APPROVED로 승인한다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "A. Auth \u0026 Device Trust"
                ],
                "summary": "기기 승인",
                "parameters": [
                    {
                        "type": "string",
                        "description": "기기 등록 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.DeviceRegistration"
                        }
                    }
                }
            }
        },
        "/device-registrations/{id}/reject": {
            "post": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "PENDING 상태인 기기를 REJECTED로 거절한다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "A. Auth \u0026 Device Trust"
                ],
                "summary": "기기 거절",
                "parameters": [
                    {
                        "type": "string",
                        "description": "기기 등록 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.DeviceRegistration"
                        }
                    }
                }
            }
        },
        "/device-registrations/{id}/revoke": {
            "post": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "APPROVED 기기의 권한을 REVOKED로 박탈한다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "A. Auth \u0026 Device Trust"
                ],
                "summary": "기기 신뢰 취소 (폐기/분실)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "기기 등록 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.DeviceRegistration"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_infrastructure_web.DeviceRegistration": {
            "type": "object",
            "properties": {
                "approvedAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "createdAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "deviceName": {
                    "type": "string",
                    "example": "iPad Pro #3"
                },
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "PENDING",
                        "APPROVED",
                        "REJECTED",
                        "REVOKED"
                    ]
                }
            }
        },
        "internal_infrastructure_web.DeviceRegistrationRequest": {
            "type": "object",
            "properties": {
                "campId": {
                    "description": "Using campId because RequestRegistration expects a campID",
                    "type": "string"
                },
                "deviceName": {
                    "type": "string"
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "ADMIN",
                        "FACILITATOR"
                    ]
                }
            }
        },
...
    }
}